backend/requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
gunicorn==22.0.0
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.12.2
whitenoise==6.7.0
